Add files via upload
This is a powershell script used to initialize and rename all of your initialized drives. It was written for a server build script to initialize and name all drives after the server was built. It renames the CD drive to (Q) so that we could keep with our naming convention and start with D on these drives.