Add files via upload
This is the command line version of an app I created to quickly reset passwords in Active Directory. It will create a password of appropriate length and complexity based on the type of account. It will then reset their password, unlock their account, update their description and set the change password on next logon to true. It sends an encrypted email to the user's email that was pulled from AD