Add files via upload
GUI version of the password reset app that changes passwords based on their level of needed security. The password requirements needed for you will have to be changed in this script. I utilized the ps2exe tool to create an executable to send to the team. You can find that here: https://github.com/rzander/PS2EXE